requirements.txt
beautifulsoup4==4.4.0
requests_html
psycopg2-binary
